1.4 rebase: Disable some tests in Spark 3.2 that fail due to HIVE-24920
HIVE-24920 is backported to 7.2.16.0 Hive. This causes some tests to fail.
For Spark 3.3, we made a change in 7.2.16.0 to account for the Hive change,
so the same tests now pass.
However, for Spark 3.2 we are using the version released in 7.2.15.0, and
as we use the same Hive dependencies throughout, the tests will still fail.
We therefore disable those tests.