Fix readable_metrics test failures due to dependence on gzip compression of parquet data.
Upstream default parquet codec is gzip, but our default is snappy.
The tests use FileHelpers.writeDataFiles to write data files directly;
we make the helper function use the codec from table properties, and
in the tests we explicitly spell out the parquet write codec to be
gzip in table properties.