1.4 rebase: Fix and enable Spark 3.3
In 1.4, the default format version is now 2. However, upstream,
copy-on-write is still used by default for v2 tables. This causes
failures in TestDeleteFrom only for Spark 3.3, because for Spark 3.3,
some support for row level operations are in Spark itself and some
need to be provided in Iceberg itself in spark-extensions. However,
TestDeleteFrom is a test in the spark module, which doesn't depend
on spark-extensions, so rules from IcebergSparkSessionsExtensions
are not applied (in fact, are not available to be applied). Thus
we need to use copy-on-write in TestDeleteFrom.